kafka: consolidate batch adapter related types
This is to avoid a circular dependency between kafka/types.h and batch
adapter when we add api_version to the batch adapter.
