+++ src/v/kafka/types.h
@@ -11,7 +11,6 @@
 
 #pragma once
 #include "bytes/bytes.h"
-#include "kafka/protocol/kafka_batch_adapter.h"
 #include "reflection/adl.h"
 #include "seastarx.h"
 #include "utils/named_type.h"
@@ -148,38 +147,6 @@ struct member_protocol {
 std::ostream& operator<<(std::ostream&, const member_protocol&);
 
 using assignments_type = std::unordered_map<member_id, bytes>;
-
-/*
- * Helper wrapper type that handles conversion when encoding/decoding produce
- * requests. This type replaces iobuf in the generated code for kafka request
- * types.
- */
-struct produce_request_record_data {
-    explicit produce_request_record_data(std::optional<iobuf>&& data) {
-        if (data) {
-            adapter.adapt(std::move(*data));
-        }
-    }
-
-    explicit produce_request_record_data(model::record_batch&& batch) {
-        adapter.v2_format = true;
-        adapter.valid_crc = true;
-        adapter.batch = std::move(batch);
-    }
-
-    kafka_batch_adapter adapter;
-};
-
-inline std::ostream&
-operator<<(std::ostream& os, const produce_request_record_data& data) {
-    fmt::print(
-      os,
-      "batch {} v2_format {} valid_crc {}",
-      data.adapter.batch ? data.adapter.batch->size_bytes() : -1,
-      data.adapter.v2_format,
-      data.adapter.valid_crc);
-    return os;
-}
 
 } // namespace kafka
